LayoutTests/platform/ios-simulator/editing/pasteboard/copy-image-with-alt-text-expected.txt

This tests that image alt text makes it to the clipboard. This test requires DRT or WTR. The first two dumps should not include alt text and the third and forth dumps should include alt text.

Dump of markup 1:
|   <shadow:root>
|     <div>
|       contenteditable="plaintext-only"

Dump of markup 2:
| 

Dump of markup 3:
|   <shadow:root>
|     <div>
|       contenteditable="plaintext-only"

Dump of markup 4:
| 
